:truck: itemHandler 통합
