+++ src/lib/msw/mocks/handlers.ts
@@ -1,6 +1,5 @@
 import authHandlers from './authHandlers'
-import { itemHandlers } from './itemHandler'
-import { testHandlers } from './testHandler'
 import { itemHandlers } from './itemHandlers'
+import { testHandlers } from './testHandler'
 
 export const handlers = [...testHandlers, ...authHandlers, ...itemHandlers]
